Refactor Buckets to extract GLSL helper
